Move QgsHistoryEntry to separate file
