+++ python/gui/auto_generated/history/qgshistoryentry.sip.in
@@ -1,0 +1,69 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/history/qgshistoryentry.h                                    *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+class QgsHistoryEntry
+{
+%Docstring(signature="appended")
+Encapsulates a history entry.
+
+.. versionadded:: 3.24
+%End
+
+%TypeHeaderCode
+#include "qgshistoryentry.h"
+%End
+  public:
+
+    QgsHistoryEntry();
+%Docstring
+Constructor for an invalid entry.
+%End
+
+    QgsHistoryEntry( const QString &providerId, const QDateTime &timestamp, const QVariantMap &entry );
+%Docstring
+Constructor for QgsHistoryEntry ``entry``, with the specified ``providerId`` and ``timestamp``.
+%End
+
+    QgsHistoryEntry( const QVariantMap &entry );
+%Docstring
+Constructor for QgsHistoryEntry ``entry``.
+
+The entry timestamp will be automatically set to the current date/time.
+%End
+
+    bool isValid() const;
+%Docstring
+Returns ``True`` if the entry is valid.
+
+.. versionadded:: 3.32
+%End
+
+    QDateTime timestamp;
+
+    QString providerId;
+
+    QVariantMap entry;
+
+    SIP_PYOBJECT __repr__();
+%MethodCode
+    const QString str = QStringLiteral( "<QgsHistoryEntry: %1 %2>" ).arg( sipCpp->providerId, sipCpp->timestamp.toString( Qt::ISODate ) );
+    sipRes = PyUnicode_FromString( str.toUtf8().constData() );
+%End
+
+};
+
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/history/qgshistoryentry.h                                    *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
