+++ python/gui/auto_generated/history/qgshistoryproviderregistry.sip.in
@@ -10,58 +10,6 @@
 
 
 
-class QgsHistoryEntry
-{
-%Docstring(signature="appended")
-Encapsulates a history entry.
-
-.. versionadded:: 3.24
-%End
-
-%TypeHeaderCode
-#include "qgshistoryproviderregistry.h"
-%End
-  public:
-
-    QgsHistoryEntry();
-%Docstring
-Constructor for an invalid entry.
-%End
-
-    QgsHistoryEntry( const QString &providerId, const QDateTime &timestamp, const QVariantMap &entry );
-%Docstring
-Constructor for QgsHistoryEntry ``entry``, with the specified ``providerId`` and ``timestamp``.
-%End
-
-    QgsHistoryEntry( const QVariantMap &entry );
-%Docstring
-Constructor for QgsHistoryEntry ``entry``.
-
-The entry timestamp will be automatically set to the current date/time.
-%End
-
-    bool isValid() const;
-%Docstring
-Returns ``True`` if the entry is valid.
-
-.. versionadded:: 3.32
-%End
-
-    QDateTime timestamp;
-
-    QString providerId;
-
-    QVariantMap entry;
-
-    SIP_PYOBJECT __repr__();
-%MethodCode
-    const QString str = QStringLiteral( "<QgsHistoryEntry: %1 %2>" ).arg( sipCpp->providerId, sipCpp->timestamp.toString( Qt::ISODate ) );
-    sipRes = PyUnicode_FromString( str.toUtf8().constData() );
-%End
-
-};
-
-
 class QgsHistoryProviderRegistry : QObject
 {
 %Docstring(signature="appended")
